Moved helper functions to separate utils module.
